LayoutTests/inspector/canvas/updateShader-webgpu-sharedVertexFragment-expected.txt

WebGPU tests for Canvas.updateShader command when the vertex and fragment shaders of a render pipeline share the same module.


== Running test suite: Canvas.updateShader.WebGPU.SharedVertexFragment
-- Running test case: Canvas.updateShader.WebGPU.SharedVertexFragment.SinglePipeline
Creating render pipeline...
PASS: Vertex and Fragment shader sources should be the same.
Updating vertex shader source...
PASS: Vertex and Fragment shader sources should be the same.
PASS: Vertex source should have changed.
Updating fragment shader source...
PASS: Vertex and Fragment shader sources should be the same.
PASS: Fragment source should have changed.

-- Running test case: Canvas.updateShader.WebGPU.SharedVertexFragment.MultiplePipelines
Creating render pipeline...
Creating render pipeline...
PASS: Both pipelines should have the same source.
Updating pipeline 1 vertex source...
PASS: Source of pipeline 1 should have changed.
PASS: Both pipelines should have different sources.
